Edited order's color UI

+++ app/src/main/java/com/example/restauyou/ModelClass/CartItem.java
@@ -26,7 +26,7 @@ public class CartItem {
     }
 
     public double getTotalPrice(){
-        return (double) quantity*(Integer.parseInt(menuItem.getItemPrice()));
+        return (double) quantity*(Double.parseDouble(menuItem.getItemPrice()));
 
     }
